nuke debug utils + rename presense to presence
